requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
requests==2.32.3
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pgeocode==0.5.0
